requirements.txt
streamlit>=1.20.0
openai>=0.27.0
pdfplumber>=0.7.6
python-docx>=0.8.11
requests>=2.28.0
